Move statement-store helpers to statement module; run zombienet on every PR
